+++ Dtos/Incomes/GeThisYearIncomes.cs
@@ -3,5 +3,5 @@
 public class GeThisYearIncomes
 {
     public List<string> Monthes { get; set; }
-    public List<double> Values { get; set; }
+    public List<decimal> Values { get; set; }
 }
